fix(eval): keep package APIs in library mode
