+++ src/tracing/evaluatorTracing.ts
@@ -4,7 +4,8 @@ import { getEnvBool } from '../envars';
 import logger from '../logger';
 import telemetry from '../telemetry';
 
-import type { EvaluateOptions, TestCase, TestSuite } from '../types/index';
+import type { TestCase, TestSuite } from '../types/index';
+import type { InternalEvaluateOptions as EvaluateOptions } from '../types/internal';
 
 // Track whether OTLP receiver has been started
 let otlpReceiverStarted = false;
