+++ src/types/internal.ts
@@ -1,0 +1,11 @@
+import type { EventSource } from './eventSource';
+import type { EvaluateOptions } from './index';
+
+/**
+ * Internal orchestration metadata that should not be accepted from reusable
+ * package callers. Process-lifecycle behavior keys off `eventSource`, so keep it
+ * separate from the public `EvaluateOptions` surface.
+ */
+export type InternalEvaluateOptions = EvaluateOptions & {
+  eventSource?: EventSource;
+};
