+++ test/server/routes/eval.sharing.test.ts
@@ -4,12 +4,10 @@ import request from 'supertest';
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest';
 
 // Mock dependencies BEFORE imports
-vi.mock('../../../src/index', () => ({
-  default: {
-    evaluate: vi.fn().mockResolvedValue({
-      toEvaluateSummary: vi.fn().mockResolvedValue({ results: [] }),
-    }),
-  },
+vi.mock('../../../src/evaluate', () => ({
+  evaluateWithSource: vi.fn().mockResolvedValue({
+    toEvaluateSummary: vi.fn().mockResolvedValue({ results: [] }),
+  }),
 }));
 
 vi.mock('../../../src/util/sharing', () => ({
@@ -23,7 +21,7 @@ vi.mock('../../../src/models/eval', () => ({
 }));
 vi.mock('../../../src/globalConfig/accounts');
 
-import promptfoo from '../../../src/index';
+import { evaluateWithSource } from '../../../src/evaluate';
 import logger from '../../../src/logger';
 import Eval from '../../../src/models/eval';
 import { createApp } from '../../../src/server/server';
@@ -31,7 +29,7 @@ import { shouldShareResults } from '../../../src/util/sharing';
 
 const errorSpy = vi.spyOn(logger, 'error');
 const mockedEvalCreate = vi.mocked(Eval.create);
-const mockedEvaluate = vi.mocked(promptfoo.evaluate);
+const mockedEvaluateWithSource = vi.mocked(evaluateWithSource);
 const mockedShouldShareResults = vi.mocked(shouldShareResults);
 
 describe('Eval Routes - Sharing behavior', () => {
@@ -60,7 +58,7 @@ describe('Eval Routes - Sharing behavior', () => {
     vi.resetAllMocks();
     errorSpy.mockClear();
 
-    mockedEvaluate.mockResolvedValue({
+    mockedEvaluateWithSource.mockResolvedValue({
       toEvaluateSummary: vi.fn().mockResolvedValue({ results: [] }),
     } as any);
 
@@ -84,10 +82,10 @@ describe('Eval Routes - Sharing behavior', () => {
 
     // Wait for async evaluate call
     await vi.waitFor(() => {
-      expect(mockedEvaluate).toHaveBeenCalled();
+      expect(mockedEvaluateWithSource).toHaveBeenCalled();
     });
 
-    const evaluateArg = mockedEvaluate.mock.calls[0][0] as any;
+    const evaluateArg = mockedEvaluateWithSource.mock.calls[0][0] as any;
     expect(evaluateArg.sharing).toBe(true);
   });
 
@@ -97,10 +95,10 @@ describe('Eval Routes - Sharing behavior', () => {
     await postJob({ ...minimalTestSuite, sharing: false });
 
     await vi.waitFor(() => {
-      expect(mockedEvaluate).toHaveBeenCalled();
+      expect(mockedEvaluateWithSource).toHaveBeenCalled();
     });
 
-    const evaluateArg = mockedEvaluate.mock.calls[0][0] as any;
+    const evaluateArg = mockedEvaluateWithSource.mock.calls[0][0] as any;
     expect(evaluateArg.sharing).toBe(false);
   });
 
@@ -110,16 +108,16 @@ describe('Eval Routes - Sharing behavior', () => {
     await postJob(minimalTestSuite);
 
     await vi.waitFor(() => {
-      expect(mockedEvaluate).toHaveBeenCalled();
+      expect(mockedEvaluateWithSource).toHaveBeenCalled();
     });
 
-    const evaluateArg = mockedEvaluate.mock.calls[0][0] as any;
+    const evaluateArg = mockedEvaluateWithSource.mock.calls[0][0] as any;
     expect(evaluateArg.sharing).toBe(true);
     expect(mockedShouldShareResults).toHaveBeenCalledWith({});
   });
 
   it('should not log the raw job body on evaluation failure', async () => {
-    mockedEvaluate.mockRejectedValueOnce(new Error('boom'));
+    mockedEvaluateWithSource.mockRejectedValueOnce(new Error('boom'));
 
     const sensitiveBody = {
       prompts: ['test prompt'],
@@ -186,10 +184,10 @@ describe('Eval Routes - Sharing behavior', () => {
     await postJob(minimalTestSuite);
 
     await vi.waitFor(() => {
-      expect(mockedEvaluate).toHaveBeenCalled();
+      expect(mockedEvaluateWithSource).toHaveBeenCalled();
     });
 
-    const evaluateArg = mockedEvaluate.mock.calls[0][0] as any;
+    const evaluateArg = mockedEvaluateWithSource.mock.calls[0][0] as any;
     expect(evaluateArg.sharing).toBe(false);
   });
 });
